Refactors validator results to allow decorators
Realized that decoration is likely the best path for altering the values
returned via `getMessages()`. To enable that, added a new interface,
`Result`, describing the capabilities of a validation result, and
modified `ValidatorResult` to implement it.

Renamed `ValidatorResultTranslator` to `TranslatableValidatorResult`; it
now accepts a `Result` to its constructor and implements `Result`. For
all methods other than `getMessages()`, it proxies to the underlying
`Result` instance (these are implemented by a new trait,
`ValidatorResultDecorator`).

Also provides a new result type, `ObscuredValueValidatorResult`; it
overrides the `getValue()` method to obscure the value, and the
`getMessages()` method to ensure that its own `getValue()` is called
when performing message template substitutions.

A new method was extracted within `ValidatorResultMessageInterpolator`,
`castValueToString()`; this was done to allow the
`ObscuredValueValidatorResult` to work with the string value when
performing obfuscation.