Begins refactoring validators, starting with Between
Adds `AbstractValidator::createInvalidResult()`, for automating creation
of the `Result` instance returned by validators; that method retrieves
method templates based on the message keys passed to it, and pulls the
message variables from the instance, passing them along with the value
to the `ValidatorResult::createInvalidResult()` constructor.

If the "value obscured" flag is enabled, decorates the result with
`ObscuredValueValidatorResult` before returning it.

The `Between` validator was refactored to use this approach.