Renames `Validator::isValid()` to `Validator::validate()`
Done for two reasons. First, the method no longer returns a boolean, so
calling it `isValid()` did not indicate that there might be a
non-boolean result. Second, having the name different will allow us to
add the `Validator` interface to v2 releases, and thus provide a
forwards-compat path for developers.