Updates Between validator to have concrete constructor arguments
Updates the `Between` validator to no longer accept an array of options,
but instead three discrete arguments. It also sets the
`$messageVariables` property during instantiation.

Tests were refactored to use data providers, and to test for both valid
arguments and the expected behavior, as well as invalid constructor
arguments and expected exceptions.